Better diagnostics.
