+++ txcache/constants.go
@@ -1,5 +1,6 @@
 package txcache
 
 const excellentGasPriceFactor = 5
-
 const maxSenderScore = 100
+const diagnosisMaxSendersToDisplay = 1000
+const diagnosisMaxTransactionsToDisplay = 10000
