+++ txcache/txByHashMap.go
@@ -74,8 +74,6 @@ func (txMap *txByHashMap) RemoveTxsBulk(txHashes [][]byte) uint32 {
 		}
 	}
 
-	logRemove.Trace("RemoveTxsBulk()", "len(txHashes)", len(txHashes), "numRemoved", numRemoved)
-
 	return numRemoved
 }
